+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - #73: Add new application metric: `Circuit Operational status` to show the circuit status depending on related Circuit Maintenances.
 
+### Fixed
+
+- #74: Fix Gmail API `after` format.
+
 ## v0.1.9 - 2021-08-12
 
 ### Changed