+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v0.1.10
 
+### Added
+
+- #73: Add new application metric: `circuit_maintenance_status` to show the circuit status depending on related Circuit Maintenances.
+
 ### Fixed
 
 - #74: Fix Gmail API `after` format.